Add Rector for PHP 8 with PR testing
This adds a Rector script to downgrade Cucumber Linter to PHP 8 which
allows us to develop with the newest PHP features on PHP 8.2 or 8.3
while installing can happen on PHP 8.

To ensure this works we add testing for pull requests on the downgraded
PHP version.

We can still improve here and use a class loader in a similar manner as
PHPStan does to load classes from within the compiled PHAR so that we
know we're actually testing the downgraded and packaged version rather
than only the downgraded code.