Ensure PHP 8.1 compatibility
We must disable the platform check from composer to make sure that the
linter will start on PHP 8.1 without complaining about PHP 8.2

This means that the burden of checking our dependencies for minimal PHP
compatibility falls with us.

We up the `phar.yml` target from 8.0 to 8.1 because cucumber/gherkin
requires PHP 8.1 in all versions. This also means it's not useful to
downgrade Symfony to try and support lower PHP versions.