Ensure linter doesn't report errors for errorless files
The way the errors were accumulated in LintCommand caused the `$errors`
array to contain empty arrays for files without errors. This caused the
error reporter to think that there are errors and exit with a non-zero
status code breaking CI.

This commit ensures that files without errors are properly excluded from
the errors array and adds a test coverage for the lint command to check
for the correct status code.