Ensure comments are allowed in between steps
We don't want to consider comments for the number of blank lines because
you might want to preface a certain block of steps by a short
explanation of a business decision. The blank line should then still be
used to visually separate blocks rather than the comments.

Because comments are not part of steps but are only in the wider
GherkinDocument this requires us to update our lint* functions and
provide it with the GherkinDocument. To make PHPStan easier to satisfy
for our not NULL validation of the feature name we pass in
`$document->uri` separately.